# development/docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Catalytic Computing SaaS API",
...
    "version": "1.0.0",
    "contact": {
      "name": "Catalytic Computing Support",
      "email": "support@catalyticcomputing.example.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://catalyticcomputing.example.com/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.catalyticcomputing.example.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Login and get JWT tokens",
        "description": "Authenticate user and receive access + refresh tokens",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "admin@acme.com",
                "password": "SecurePassword123!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
    "/api/lattices": {
      "post": {
        "tags": [
          "lattices"
        ],
        "summary": "Create new KA lattice",
        "description": "Create a new Knowledge Algebra lattice with specified dimensions",
        "operationId": "createLattice",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LatticeCreate"
              },
              "example": {
                "name": "My Lattice",
                "dimensions": 3,
                "size": 1000,
                "field_type": "complex",
                "geometry": "euclidean",
                "enable_gpu": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lattice created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LatticeResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitExceeded"
          }
        }
      },
      "get": {
        "tags": [
          "lattices"
        ],
        "summary": "List tenant lattices",
        "description": "Get all lattices for current tenant",
        "operationId": "listLattices",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of lattices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LatticeResponse"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from /api/auth/login"
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "example": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "description": "Seconds until expiration"
          }
        }
      },
...
      "LatticeCreate": {
        "type": "object",
        "required": [
          "name",
          "dimensions",
          "size"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "dimensions": {
            "type": "integer",
            "minimum": 2,
            "maximum": 10
          },
          "size": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10000
          },
          "field_type": {
            "type": "string",
            "enum": [
              "real",
              "complex"
            ],
            "default": "complex"
          },
          "geometry": {
            "type": "string",
            "enum": [
              "euclidean",
              "hyperbolic",
              "spherical"
            ],
            "default": "euclidean"
          },
          "enable_gpu": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "LatticeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "dimensions": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          },
          "field_type": {
            "type": "string"
          },
          "geometry": {
            "type": "string"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "processing_info": {
            "type": "object",
            "nullable": true,
            "properties": {
              "gpu_used": {
                "type": "boolean"
              },
              "processing_time_ms": {
                "type": "number"
              }
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          },
          "status_code": {
            "type": "integer"
          },
          "type": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request - invalid input",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - missing or invalid authentication",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "RateLimitExceeded": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}